Palettes: Added presets., export to .pal, memory of last  5 .pal files used

+++ GUI.NET/InteropEmu.cs
@@ -117,7 +117,7 @@ namespace Mesen.GUI
 		[DllImport(DLLPath)] public static extern void SetVideoAspectRatio(VideoAspectRatio aspectRatio);
 		[DllImport(DLLPath)] public static extern void SetVideoFilter(VideoFilterType filter);
 		[DllImport(DLLPath)] public static extern void SetVideoResizeFilter(VideoResizeFilter filter);
-		[DllImport(DLLPath)] public static extern void SetRgbPalette(Int32[] palette);
+		[DllImport(DLLPath)] public static extern void SetRgbPalette(byte[] palette);
 		[DllImport(DLLPath)] public static extern void SetPictureSettings(double brightness, double contrast, double saturation, double hue, double scanlineIntensity);
 		[DllImport(DLLPath)] public static extern void SetNtscFilterSettings(double artifacts, double bleed, double fringing, double gamma, double resolution, double sharpness, [MarshalAs(UnmanagedType.I1)]bool mergeFields);
